Renamed some files and properties
